Change:ダイアログを出すコードを別ファイルに移動
